refactor(discovery): minmize changes to focus on Daydream for orchestrator discovery plans
- Replaced PymtHouse manifest dependencies with Daydream-specific logic in the capability catalog and Python gateway routes.
- Updated billing provider handling to default to 'daydream' instead of 'pymthouse'.
- Introduced new discovery constants and policies to support Daydream-only functionality.
- Removed legacy PymtHouse manifest and discovery plan files, streamlining the codebase.
- Enhanced the orchestrator discovery policy to apply filters without relying on PymtHouse exclusions.
