+++ apps/web-next/src/app/api/v1/orchestrator-leaderboard/dataset/refresh/route.ts
@@ -17,8 +17,6 @@ import { validateSession } from '@/lib/api/auth';
 import { getAuthToken } from '@/lib/api/response';
 import { getRefreshIntervalMs, getLastRefreshedAt } from '@/lib/orchestrator-leaderboard/config';
 import { refreshGlobalDataset } from '@/lib/orchestrator-leaderboard/global-refresh';
-import { syncPymthouseManifestSnapshot } from '@/lib/pymthouse-manifest';
-import { clearPlanCache } from '@/lib/orchestrator-leaderboard/refresh';
 
 function isCronAuth(request: NextRequest): boolean {
   const secret = process.env.CRON_SECRET;
@@ -61,17 +59,12 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
   if (cronAuthed) {
     const lastRefreshed = await getLastRefreshedAt();
     if (lastRefreshed && Date.now() - lastRefreshed.getTime() < intervalMs) {
-      const { revisionChanged } = await syncPymthouseManifestSnapshot();
-      if (revisionChanged) {
-        clearPlanCache();
-      }
       return NextResponse.json({
         success: true,
         data: {
           skipped: true,
           reason: 'Global dataset is still fresh',
           nextRefreshInMs: intervalMs - (Date.now() - lastRefreshed.getTime()),
-          pymthouseManifestRevisionChanged: revisionChanged,
         },
       });
     }
