+++ apps/web-next/src/app/api/v1/orchestrator-leaderboard/python-gateway/route.ts
@@ -18,10 +18,7 @@ import {
   isCapabilityAllowedForProvider,
   normalizeBillingProviderSlug,
 } from '@/lib/orchestrator-leaderboard/provider-restrictions';
-import {
-  DISCOVERY_RESPONSE_CACHE_CONTROL,
-  ensurePymthouseManifestFresh,
-} from '@/lib/pymthouse-manifest';
+import { DISCOVERY_RESPONSE_CACHE_CONTROL } from '@/lib/orchestrator-leaderboard/discovery-constants';
 
 const DEFAULT_CAPABILITY = 'noop';
 const DEFAULT_TOP_N = 100;
@@ -85,10 +82,6 @@ export async function GET(request: NextRequest): Promise<Response> {
   const authToken = getAuthToken(request) || '';
 
   try {
-    if (billingProvider === 'pymthouse') {
-      await ensurePymthouseManifestFresh();
-    }
-
     const ordered: string[] = [];
     const seen = new Set<string>();
     let cacheAgeMs = 0;
