+++ apps/web-next/src/lib/orchestrator-leaderboard/__tests__/provider-restrictions.test.ts
@@ -1,9 +1,5 @@
 import { afterEach, describe, expect, it } from 'vitest';
 import {
-  resetPymthouseManifestCacheForTests,
-  seedPymthouseManifestForTests,
-} from '@/lib/pymthouse-manifest';
-import {
   filterCapabilitiesForProvider,
   isCapabilityAllowedForProvider,
   normalizeBillingProviderSlug,
@@ -13,69 +9,44 @@ import {
 
 describe('provider-restrictions', () => {
   afterEach(() => {
-    resetPymthouseManifestCacheForTests();
+    // Daydream-only: no manifest snapshot state to reset.
   });
 
   it('normalizes supported slugs and rejects unknown values', () => {
-    expect(normalizeBillingProviderSlug(' PYMTHOUSE ')).toBe('pymthouse');
+    expect(normalizeBillingProviderSlug(' PYMTHOUSE ')).toBeNull();
     expect(normalizeBillingProviderSlug('daydream')).toBe('daydream');
     expect(normalizeBillingProviderSlug('stripe')).toBeNull();
     expect(normalizeBillingProviderSlug(null)).toBeNull();
   });
 
-  it('uses pymthouse manifest revision when provider is pymthouse', () => {
-    seedPymthouseManifestForTests({
-      capabilities: [{ pipeline: 'video', modelId: 'model-a' }],
-      excludedCapabilities: [],
-    });
-
-    expect(providerRestrictionRevision('pymthouse')).not.toBe('na');
+  it('uses a stable restriction revision for caching', () => {
     expect(providerRestrictionRevision('daydream')).toBe('na');
+    expect(providerRestrictionRevision('pymthouse')).toBe('na');
   });
 
-  it('filters capabilities by provider exclusion rules', () => {
-    seedPymthouseManifestForTests({
-      capabilities: [{ pipeline: 'video', modelId: 'model-a' }],
-      excludedCapabilities: [{ pipeline: 'video', modelId: 'model-b' }],
-    });
-
+  it('does not filter capabilities for Daydream', () => {
     const caps = ['video/model-a', 'video/model-b', 'future/model-c'];
-    expect(filterCapabilitiesForProvider(caps, 'pymthouse')).toEqual(['video/model-a', 'future/model-c']);
+    expect(filterCapabilitiesForProvider(caps, 'daydream')).toEqual(caps);
     expect(filterCapabilitiesForProvider(caps, 'daydream')).toEqual(caps);
   });
 
-  it('resolves plan capabilities with provider-aware filtering', () => {
-    seedPymthouseManifestForTests({
-      capabilities: [{ pipeline: 'video', modelId: 'model-a' }],
-      excludedCapabilities: [{ pipeline: 'video', modelId: 'model-z' }],
-    });
-
+  it('resolves plan capabilities without manifest filtering', () => {
     expect(resolvePlanCapabilitiesForProvider({
-      billingProviderSlug: 'pymthouse',
+      billingProviderSlug: 'daydream',
       capabilities: ['video/model-a', 'video/model-z', 'future/model-c'],
-    } as const)).toEqual(['video/model-a', 'future/model-c']);
+    } as const)).toEqual(['video/model-a', 'video/model-z', 'future/model-c']);
   });
 
-  it('does not apply manifest filter when billingProviderSlug is null (backfill target)', () => {
-    seedPymthouseManifestForTests({
-      capabilities: [{ pipeline: 'video', modelId: 'model-a' }],
-      excludedCapabilities: [{ pipeline: 'video', modelId: 'model-b' }],
-    });
-
+  it('does not filter capabilities when billingProviderSlug is null', () => {
     const caps = ['video/model-a', 'video/model-b'];
     expect(filterCapabilitiesForProvider(caps, null)).toEqual(caps);
-    expect(filterCapabilitiesForProvider(caps, 'pymthouse')).toEqual(['video/model-a']);
+    expect(filterCapabilitiesForProvider(caps, 'pymthouse')).toEqual(caps);
   });
 
-  it('checks capability allow decision via provider semantics', () => {
-    seedPymthouseManifestForTests({
-      capabilities: [{ pipeline: 'video', modelId: 'model-a' }],
-      excludedCapabilities: [{ pipeline: 'video', modelId: 'model-b' }],
-    });
-
-    expect(isCapabilityAllowedForProvider('video/model-a', 'pymthouse')).toBe(true);
-    expect(isCapabilityAllowedForProvider('video/model-b', 'pymthouse')).toBe(false);
-    expect(isCapabilityAllowedForProvider('future/model-c', 'pymthouse')).toBe(true);
+  it('allows capabilities under Daydream semantics', () => {
+    expect(isCapabilityAllowedForProvider('video/model-a', 'daydream')).toBe(true);
     expect(isCapabilityAllowedForProvider('video/model-b', 'daydream')).toBe(true);
+    // Unknown/invalid providers do not deny capabilities in PR #337.
+    expect(isCapabilityAllowedForProvider('video/model-b', 'pymthouse')).toBe(true);
   });
 });
