--- apps/web-next/src/lib/pymthouse-discovery-plans.test.ts
@@ -1,87 +1,0 @@
-/** @vitest-environment node */
-
-import { afterEach, describe, expect, it, vi } from 'vitest';
-
-import {
-  fetchPymthouseDiscoveryPlans,
-  getPymthouseApiV1Base,
-  mapPymthousePlansResponse,
-  mergeDiscoveryPolicies,
-  resetPymthouseDiscoveryPlansCacheForTests,
-} from '@/lib/pymthouse-discovery-plans';
-
-describe('pymthouse-discovery-plans', () => {
-  afterEach(() => {
-    vi.unstubAllEnvs();
-    resetPymthouseDiscoveryPlansCacheForTests();
-  });
-
-  it('getPymthouseApiV1Base strips /oidc suffix', () => {
-    vi.stubEnv('PYMTHOUSE_ISSUER_URL', 'https://ph.example/api/v1/oidc/');
-    expect(getPymthouseApiV1Base()).toBe('https://ph.example/api/v1');
-  });
-
-  it('mergeDiscoveryPolicies uses min topN', () => {
-    expect(mergeDiscoveryPolicies({ topN: 5 }, { topN: 20 })?.topN).toBe(5);
-    expect(mergeDiscoveryPolicies({ topN: 20 }, { topN: 5 })?.topN).toBe(5);
-  });
-
-  it('fetchPymthouseDiscoveryPlans returns null without env', async () => {
-    expect(await fetchPymthouseDiscoveryPlans({ skipCache: true })).toBeNull();
-  });
-
-  it('fetchPymthouseDiscoveryPlans parses JSON on 200', async () => {
-    vi.stubEnv('PYMTHOUSE_ISSUER_URL', 'http://localhost:3001/api/v1/oidc');
-    vi.stubEnv('PYMTHOUSE_PUBLIC_CLIENT_ID', 'app_x');
-    vi.stubEnv('PYMTHOUSE_M2M_CLIENT_ID', 'm2m_x');
-    vi.stubEnv('PYMTHOUSE_M2M_CLIENT_SECRET', 'secret');
-
-    const fetchMock = vi.fn(async () =>
-      ({
-        ok: true,
-        json: async () => ({
-          plans: [{ id: 'p1', name: 'P', status: 'active', discoveryPolicy: null, capabilities: [] }],
-        }),
-      } as Response),
-    );
-    vi.stubGlobal('fetch', fetchMock);
-
-    const out = await fetchPymthouseDiscoveryPlans({ skipCache: true });
-    expect(out?.plans).toHaveLength(1);
-    expect(out?.plans[0].id).toBe('p1');
-    expect(String(fetchMock.mock.calls[0]?.[0])).toContain('/apps/app_x/plans');
-    expect(String(fetchMock.mock.calls[0]?.[0])).not.toContain('discovery');
-  });
-
-  it('mapPymthousePlansResponse drops network default and inactive', () => {
-    const raw = {
-      plans: [
-        {
-          id: 'net',
-          name: '__pymthouse_network_default__',
-          status: 'active',
-          isNetworkDefault: true,
-          discoveryPolicy: null,
-          capabilities: [],
-        },
-        { id: 'p1', name: 'Custom', status: 'draft', discoveryPolicy: null, capabilities: [] },
-        {
-          id: 'p2',
-          name: 'Live',
-          status: 'active',
-          isNetworkDefault: false,
-          discoveryPolicy: { topN: 3 },
-          capabilities: [{ pipeline: 'llm', modelId: 'm1', discoveryPolicy: null }],
-        },
-      ],
-    };
-    const out = mapPymthousePlansResponse(raw);
-    expect(out?.plans).toHaveLength(1);
-    expect(out?.plans[0].id).toBe('p2');
-    expect(out?.plans[0].capabilities[0]).toEqual({
-      pipeline: 'llm',
-      modelId: 'm1',
-      discoveryPolicy: null,
-    });
-  });
-});
